pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.95.0",
    "mcp>=1.27.0",
    "python-dotenv>=1.2.2",
    "playwright>=1.44.0",
    "openai>=1.79.0",
]
